+++ frontend/src/framework/components/MultiEnsembleSelect/index.ts
@@ -1,0 +1,1 @@
+export { MultiEnsembleSelect } from "./multiEnsembleSelect";
