+++ frontend/src/framework/components/SingleEnsembleSelect/index.ts
@@ -1,0 +1,1 @@
+export { SingleEnsembleSelect } from "./singleEnsembleSelect";
