+++ frontend/src/modules/Map/MapSettings.tsx
@@ -3,11 +3,11 @@ import React from "react";
 import { DynamicSurfaceDirectory, StaticSurfaceDirectory } from "@api";
 import { SurfaceStatisticFunction } from "@api";
 import { EnsembleIdent } from "@framework/EnsembleIdent";
-import { SingleEnsembleSelect } from "@framework/EnsembleSetUiComponents";
 import { fixupEnsembleIdent, maybeAssignFirstSyncedEnsemble } from "@framework/EnsembleSetUiHelpers";
 import { ModuleFCProps } from "@framework/Module";
 import { SyncSettingKey, SyncSettingsHelper } from "@framework/SyncSettings";
 import { useEnsembleSet } from "@framework/WorkbenchSession";
+import { SingleEnsembleSelect } from "@framework/components/SingleEnsembleSelect";
 import { ApiStateWrapper } from "@lib/components/ApiStateWrapper";
 import { Checkbox } from "@lib/components/Checkbox";
 import { CircularProgress } from "@lib/components/CircularProgress";
